# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FullNormies Sprite API",
    "version": "1.0.0",
    "description": "Live server-side sprite generation for Normie NFTs (token IDs 0–9999).\n\n## Native resolution\n- **40 × 80 px** portrait canvas (same aspect ratio as Normie head art)\n- Anchor: **`{ x: 20, y: … }`** from `full-meta.json` / `sheet.json` — bottom-center of feet in **stand** pose (native pixels). **`y` is per normie** (torso height + leg-length seed), usually high-50s to low-60s.\n- Place sprites on a floor by aligning `anchor.y` to the floor y-coordinate\n\n## Color palette\n- Dark: `#48494b` (charcoal)\n- Light: `#e3e5e4` (off-white)\n- Background: transparent (`PNG` with alpha)\n\n## Facing\nAll sprites face **right**. Flip horizontally in canvas/CSS for left-facing:\n- Canvas: `ctx.scale(-1, 1)` then draw at negative x offset\n- CSS: `transform: scaleX(-1)`\n\n## CORS\nAll `/api/v1/*` endpoints return `Access-Control-Allow-Origin: *`.\nSafe to call directly from any browser origin.\n\n## Caching\n- Sprites: `Cache-Control: public, max-age=3600, stale-while-revalidate=86400`\n- ETags included for cheap revalidation\n- Normie data upstream revalidates every hour from `api.normies.art`"
  },
  "servers": [
    { "url": "/api", "description": "Same-origin" },
    { "url": "https://fullnormies.vercel.app/api", "description": "Production" }
  ],
...
  "paths": {
...
    "/v1/normies/full-meta": {
      "post": {
        "tags": ["batch"],
        "summary": "Batch fetch sprite metadata",
        "description": "Fetch full-meta for up to 50 normies in one request. Ideal for prefetching a dorm roster. Returns an array in the same order as the requested IDs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "integer", "minimum": 0, "maximum": 9999 },
                    "maxItems": 50,
                    "description": "Array of normie token IDs"
                  }
                }
              },
              "example": { "ids": [1, 42, 627, 9999] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of meta objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FullMeta" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
      "Anchor": {
        "type": "object",
        "description": "Bottom-center of feet in native pixels. Use to place the sprite flush with a floor.",
        "properties": {
          "x": { "type": "integer", "description": "Horizontal anchor — always half of native width (20 for 40px sprites)", "example": 20 },
          "y": { "type": "integer", "description": "Vertical anchor — bottom of shoes in stand pose; varies per normie (typically 56–63)", "example": 60 }
        }
      },
      "FullMeta": {
        "type": "object",
        "properties": {
          "id":             { "type": "integer" },
          "exists":         { "type": "boolean" },
          "pixelWidth":     { "type": "integer", "description": "Native sprite width in pixels", "example": 40 },
          "pixelHeight":    { "type": "integer", "description": "Native sprite height in pixels", "example": 80 },
          "anchor":         { "$ref": "#/components/schemas/Anchor" },
          "posesAvailable": { "type": "array", "items": { "type": "string" } },
          "walkFrames":     { "type": "integer", "example": 4 },
          "facing":         { "type": "string", "enum": ["right"], "description": "All sprites face right; flip in client for left-facing" },
          "source":         { "type": "string" },
          "updatedAt":      { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } } } } }
      },
...
    }
  }
}